// DeltafinaWebApp/ClientApp/src/app/models/portal-farm.models.ts
export class PortalFarmDetailModel {
  id: string = "";
  code: string = "";
  name: string = "";
  taxCode: string = "";
  vat: string = "";
  email: string = "";
  phone: string = "";
  mobile: string = "";
  fax: string = "";
  url: string = "";
  street: string = "";
  streetNumber: string = "";
  addressCode: string = "";
  location: string = "";
  city: string = "";
  state: string = "";
  note: string
}

export class PortalFarmModel {
  id: string = "";
  code: string = "";
  name: string = "";
  taxCode: string = "";
  vat: string = "";
  email: string = "";
  phone: string = "";
  mobile: string = "";
  city: string = "";
}